Flisol 2025/requirements_1.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-chroma
python-dotenv
glob2

